requirements.txt
pyaudio
python-sounddevice
openai-whisper
openai
python-dotenv
